move websocket routes out of cli file
