+++ gdbgui/cli.py
@@ -12,283 +12,21 @@ import os
 import platform
 import re
 import shlex
-import traceback
-from typing import Dict, List, Optional
-
-from flask import abort, request, session
-from flask_socketio import SocketIO, emit  # type: ignore
+from typing import List, Optional
 
 from gdbgui import __version__
 
-from .server.app import app
+from .server.app import app, socketio
 from .server.constants import (
     DEFAULT_GDB_EXECUTABLE,
     DEFAULT_HOST,
     DEFAULT_PORT,
 )
-from .server.http_util import is_cross_origin
 from .server.server import run_server
-from .server.sessionmanager import DebugSession, SessionManager
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.WARNING)
 logging.basicConfig(format="(%(asctime)s) %(msg)s")
-
-
-socketio = SocketIO(manage_session=False)
-manager: SessionManager = app.config.get("_manager")  # type: ignore
-
-
-@socketio.on("connect", namespace="/gdb_listener")
-def client_connected():
-    """Connect a websocket client to a debug session
-
-    This is the main intial connection.
-
-    Depending on the arguments passed, the client will connect
-    to an existing debug session, or create a new one.
-    A message is a emitted back to the client with details on
-    the debug session that was created or connected to.
-    """
-    if is_cross_origin(request):
-        logger.warning("Received cross origin request. Aborting")
-        abort(403)
-
-    csrf_token = request.args.get("csrf_token")
-    if csrf_token is None:
-        logger.warning("Recieved invalid csrf token")
-        emit("server_error", {"message": "Recieved invalid csrf token"})
-        return
-
-    elif csrf_token != session.get("csrf_token"):
-        # this can happen fairly often, so log debug message, not warning
-        logger.debug(
-            "Recieved invalid csrf token %s (expected %s)"
-            % (csrf_token, str(session.get("csrf_token")))
-        )
-        emit(
-            "server_error", {"message": "Session expired. Please refresh this webpage."}
-        )
-        return
-
-    desired_gdbpid = int(request.args.get("gdbpid", 0))
-    try:
-        if desired_gdbpid:
-            # connect to exiting debug session
-            debug_session = manager.connect_client_to_debug_session(
-                desired_gdbpid=desired_gdbpid, client_id=request.sid
-            )
-            emit(
-                "debug_session_connection_event",
-                {
-                    "ok": True,
-                    "started_new_gdb_process": False,
-                    "pid": debug_session.pid,
-                    "message": f"Connected to existing gdb process {desired_gdbpid}",
-                },
-            )
-        else:
-            # start new debug session
-            gdb_command = request.args.get("gdb_command", app.config["gdb_command"])
-            mi_version = request.args.get("mi_version", "mi2")
-            debug_session = manager.add_new_debug_session(
-                gdb_command=gdb_command, mi_version=mi_version, client_id=request.sid
-            )
-            emit(
-                "debug_session_connection_event",
-                {
-                    "ok": True,
-                    "started_new_gdb_process": True,
-                    "message": f"Started new gdb process, pid {debug_session.pid}",
-                    "pid": debug_session.pid,
-                },
-            )
-    except Exception as e:
-        emit(
-            "debug_session_connection_event",
-            {"message": f"Failed to establish gdb session: {e}", "ok": False},
-        )
-
-    # Make sure there is a reader thread reading. One thread reads all instances.
-    if manager.gdb_reader_thread is None:
-        manager.gdb_reader_thread = socketio.start_background_task(
-            target=read_and_forward_gdb_and_pty_output
-        )
-        logger.info("Created background thread to read gdb responses")
-
-
-@socketio.on("pty_interaction", namespace="/gdb_listener")
-def pty_interaction(message):
-    """Write a character to the user facing pty"""
-    debug_session = manager.debug_session_from_client_id(request.sid)
-    if not debug_session:
-        emit(
-            "error_running_gdb_command",
-            {"message": f"no gdb session available for client id {request.sid}"},
-        )
-        return
-
-    try:
-        data = message.get("data")
-        pty_name = data.get("pty_name")
-        if pty_name == "user_pty":
-            pty = debug_session.pty_for_gdb
-        elif pty_name == "program_pty":
-            pty = debug_session.pty_for_debugged_program
-        else:
-            raise ValueError(f"Unknown pty: {pty_name}")
-
-        action = data.get("action")
-        if action == "write":
-            key = data["key"]
-            pty.write(key)
-        elif action == "set_winsize":
-            pty.set_winsize(data["rows"], data["cols"])
-        else:
-            raise ValueError(f"Unknown action {action}")
-    except Exception:
-        err = traceback.format_exc()
-        logger.error(err)
-        emit("error_running_gdb_command", {"message": err})
-
-
-@socketio.on("run_gdb_command", namespace="/gdb_listener")
-def run_gdb_command(message: Dict[str, str]):
-    """Write commands to gdbgui's gdb mi pty"""
-    client_id = request.sid  # type: ignore
-    debug_session = manager.debug_session_from_client_id(client_id)
-    if not debug_session:
-        emit("error_running_gdb_command", {"message": "no session"})
-        return
-    pty_mi = debug_session.pygdbmi_controller
-    if pty_mi is not None:
-        try:
-            # the command (string) or commands (list) to run
-            cmds = message["cmd"]
-            for cmd in cmds:
-                pty_mi.write(
-                    cmd + "\n",
-                    timeout_sec=0,
-                    raise_error_on_timeout=False,
-                    read_response=False,
-                )
-
-        except Exception:
-            err = traceback.format_exc()
-            logger.error(err)
-            emit("error_running_gdb_command", {"message": err})
-    else:
-        emit("error_running_gdb_command", {"message": "gdb is not running"})
-
-
-def send_msg_to_clients(client_ids, msg, error=False):
-    """Send message to all clients"""
-    if error:
-        stream = "stderr"
-    else:
-        stream = "stdout"
-
-    response = [{"message": None, "type": "console", "payload": msg, "stream": stream}]
-
-    for client_id in client_ids:
-        logger.info("emiting message to websocket client id " + client_id)
-        socketio.emit(
-            "gdb_response", response, namespace="/gdb_listener", room=client_id
-        )
-
-
-@socketio.on("disconnect", namespace="/gdb_listener")
-def client_disconnected():
-    """do nothing if client disconnects"""
-    manager.disconnect_client(request.sid)
-    logger.info("Client websocket disconnected, id %s" % (request.sid))
-
-
-@socketio.on("Client disconnected")
-def test_disconnect():
-    print("Client websocket disconnected", request.sid)
-
-
-def read_and_forward_gdb_and_pty_output():
-    """A task that runs on a different thread, and emits websocket messages
-    of gdb responses"""
-
-    while True:
-        socketio.sleep(0.05)
-        debug_sessions_to_remove = []
-        for debug_session, client_ids in manager.debug_session_to_client_ids.items():
-            try:
-                try:
-                    response = debug_session.pygdbmi_controller.get_gdb_response(
-                        timeout_sec=0, raise_error_on_timeout=False
-                    )
-
-                except Exception:
-                    response = None
-                    send_msg_to_clients(
-                        client_ids,
-                        "The underlying gdb process has been killed. This tab will no longer function as expected.",
-                        error=True,
-                    )
-                    debug_sessions_to_remove.append(debug_session)
-
-                if response:
-                    for client_id in client_ids:
-                        logger.info(
-                            "emiting message to websocket client id " + client_id
-                        )
-                        socketio.emit(
-                            "gdb_response",
-                            response,
-                            namespace="/gdb_listener",
-                            room=client_id,
-                        )
-                else:
-                    # there was no queued response from gdb, not a problem
-                    pass
-
-            except Exception:
-                logger.error(traceback.format_exc())
-
-        debug_sessions_to_remove += check_and_forward_pty_output()
-        for debug_session in set(debug_sessions_to_remove):
-            manager.remove_debug_session(debug_session)
-
-
-def check_and_forward_pty_output() -> List[DebugSession]:
-    debug_sessions_to_remove = []
-    for debug_session, client_ids in manager.debug_session_to_client_ids.items():
-        try:
-            response = debug_session.pty_for_gdb.read()
-            if response is not None:
-                for client_id in client_ids:
-                    socketio.emit(
-                        "user_pty_response",
-                        response,
-                        namespace="/gdb_listener",
-                        room=client_id,
-                    )
-
-            response = debug_session.pty_for_debugged_program.read()
-            if response is not None:
-                for client_id in client_ids:
-                    socketio.emit(
-                        "program_pty_response",
-                        response,
-                        namespace="/gdb_listener",
-                        room=client_id,
-                    )
-        except Exception as e:
-            debug_sessions_to_remove.append(debug_session)
-            for client_id in client_ids:
-                socketio.emit(
-                    "fatal_server_error",
-                    {"message": str(e)},
-                    namespace="/gdb_listener",
-                    room=client_id,
-                )
-            logger.error(e, exc_info=True)
-    return debug_sessions_to_remove
 
 
 def get_gdbgui_auth_user_credentials(auth_file, user, password):
